+++ tencentcloud/sms/v20210111/models.go
@@ -1371,7 +1371,7 @@ type SmsPackagesStatisticsRequest struct {
 	BeginTime *string `json:"BeginTime,omitempty" name:"BeginTime"`
 
 	// 结束时间，格式为yyyymmddhh，精确到小时，例如2021050118，表示2021年5月1号18时。
-	// 注：EndTime 必须大于 BeginTime，拉取套餐包的创建时间不大于结束时间。
+	// 注：EndTime 必须大于 BeginTime且小于当前时间，拉取套餐包的创建时间不大于结束时间。
 	EndTime *string `json:"EndTime,omitempty" name:"EndTime"`
 }