+++ tencentcloud/tdmq/v20200217/models.go
@@ -5763,7 +5763,7 @@ type ModifyCmqQueueAttributeRequest struct {
 	// 消息可见性超时。取值范围 1-43200 秒（即12小时内），默认值 30。
 	VisibilityTimeout *uint64 `json:"VisibilityTimeout,omitempty" name:"VisibilityTimeout"`
 
-	// 消息最大长度。取值范围 1024-65536 Byte（即1-64K），默认值 65536。
+	// 消息最大长度，新版CMQ新建的队列默认1024KB，不支持修改
 	MaxMsgSize *uint64 `json:"MaxMsgSize,omitempty" name:"MaxMsgSize"`
 
 	// 消息最长未确认时间。取值范围 30-43200 秒（30秒~12小时），默认值 3600 (1 小时)。