+++ tencentcloud/vpc/v20170312/client.go
@@ -1411,6 +1411,8 @@ func NewAttachNetworkInterfaceResponse() (response *AttachNetworkInterfaceRespon
 // AttachNetworkInterface
 // 本接口（AttachNetworkInterface）用于弹性网卡绑定云服务器。
 //
+// * 一个弹性网卡请至少绑定一个安全组，如需绑定请参见<a href="https://cloud.tencent.com/document/product/215/43132">弹性网卡绑定安全组</a>。
+//
 // * 一个云服务器可以绑定多个弹性网卡，但只能绑定一个主网卡。更多限制信息详见<a href="https://cloud.tencent.com/document/product/576/18527">弹性网卡使用限制</a>。
 //
 // * 一个弹性网卡只能同时绑定一个云服务器。
@@ -1439,6 +1441,8 @@ func (c *Client) AttachNetworkInterface(request *AttachNetworkInterfaceRequest) 
 
 // AttachNetworkInterface
 // 本接口（AttachNetworkInterface）用于弹性网卡绑定云服务器。
+//
+// * 一个弹性网卡请至少绑定一个安全组，如需绑定请参见<a href="https://cloud.tencent.com/document/product/215/43132">弹性网卡绑定安全组</a>。
 //
 // * 一个云服务器可以绑定多个弹性网卡，但只能绑定一个主网卡。更多限制信息详见<a href="https://cloud.tencent.com/document/product/576/18527">弹性网卡使用限制</a>。
 //